Move vSphere CAPI yaml template to separate file (#3660)
No logical changes. Just to make it easier to navigate this code.
